feat-142: BottomNavigation Fragment인지 판별하는 로직 viewModel로 옮김
