+++ presentation/src/main/java/com/plub/presentation/event/MainEvent.kt
@@ -6,4 +6,5 @@ import com.plub.domain.model.vo.signUp.hobbies.SignUpHobbiesVo
 
 sealed class MainEvent : Event {
     data class ShowBottomNavigationBadge(val index:Int) : MainEvent()
+    data class BottomNavigationVisibility(val isVisible:Boolean) : MainEvent()
 }
